backend/requirements.txt
Django==5.1.6
django-ninja==1.3.0
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
PyJWT==2.10.1
python-dotenv==1.0.1
